Updation in Acadmic module

+++ CampusTouch.Application/Features/Program/Commands/DeleteCourseHandler.cs
@@ -1,4 +1,5 @@
-﻿using CampusTouch.Application.Interfaces;
+﻿using CampusTouch.Application.Common.Exceptions;
+using CampusTouch.Application.Interfaces;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -11,14 +12,27 @@ namespace CampusTouch.Application.Features.Program.Commands
     public class DeleteCourseHandler : IRequestHandler<DeleteCourseCommand, int>
     {
         private readonly IProgramRepository _programRepository;
+        private readonly ICurrentUserService _currentUserService;
 
-        public DeleteCourseHandler(IProgramRepository programRepository)
+        public DeleteCourseHandler(IProgramRepository programRepository,ICurrentUserService currentUserService)
         {
             _programRepository = programRepository;
+            _currentUserService = currentUserService;
         }
 
         public async Task<int> Handle(DeleteCourseCommand request, CancellationToken cancellationToken)
         {
+
+            var userid = _currentUserService.UserId;
+            var ProgramIsExist = await _programRepository.GetByIdAsync(request.Id);
+            if (ProgramIsExist == null)
+               throw new NotFoundException("Program is not Exist");
+            if (!_currentUserService.IsAdmin)
+                throw new UnauthorizedException("Only Admin can Delete the Programs");
+            ProgramIsExist.IsActive=false;
+            ProgramIsExist.IsDeleted = true;
+            ProgramIsExist.DeletedAt = DateTime.UtcNow;
+            ProgramIsExist.DeletedBy=userid;
             return await _programRepository.DeleteAsync(request.Id);
         }
     }
